Feat: 모달 배경 부분 클릭하면 모달 사라
