+++ src/app/my/editInfo/page.tsx
@@ -1,13 +1,9 @@
 "use client";
-import TopNav from "@/components/common/TopNav";
 import UserInfoList from "@/components/myPage/editInfoPage/UserInfoList";
 import CheckPassword from "@/components/myPage/editInfoPage/CheckPassword";
 import userInfo from "@/mocks/seon/userInfo.json";
 import { useState } from "react";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-import { IUserEditableInfo } from "@/types/my";
-import { AxiosError } from "axios";
-import { getMyInfo } from "@/app/apis/services/auth";
+import { useQueryClient } from "@tanstack/react-query";
 
 function EditInfoPage() {
   const queryClient = useQueryClient();
