+++ src/app/page.tsx
@@ -3,7 +3,6 @@ import AdminButtonSection from "@/components/mainPage/AdminButtonSection";
 import CarouselBanner from "@/components/mainPage/CarouselBanner";
 import CustomListSection from "@/components/mainPage/CustomListSection";
 import PbListSection from "@/components/mainPage/PbListSection";
-import SearchLocation from "@/components/mainPage/SearchLocation";
 import SpecialtySection from "@/components/mainPage/SpecialtySection";
 import StockFirmSection from "@/components/mainPage/StockFirmSection";
