+++ src/components/common/Main.tsx
@@ -1,6 +1,6 @@
 "use client";
-import { getCookie, removeCookie } from "@/utils/cookies";
-import { redirect, usePathname } from "next/navigation";
+import { removeCookie } from "@/utils/cookies";
+import { usePathname } from "next/navigation";
 import { ReactNode, useEffect } from "react";
 
 function Main({ children }: { children: ReactNode }) {
