+++ src/hooks/usePBListQueries.ts
@@ -1,9 +1,8 @@
 import { createQueryString } from "@/utils/createQueryString";
-import { redirect, usePathname, useRouter, useSearchParams } from "next/navigation";
+import { redirect, useRouter, useSearchParams } from "next/navigation";
 import { MouseEvent } from "react";
 
 export const usePBListQueries = () => {
-  const pathName = usePathname();
   const searchParams = useSearchParams();
   const router = useRouter();
